Fix serialization of FileTarget
Serde's derive macros do not support newtype enum variants within an
enum with tagged representation, so serializing a FileTarget would
always fail. Thankfully, this representation wasn't intended, but
unfortunately Serde also does not support partially tagged enum
representations, such as the desired representation of FileTarget where
Automatic is simply a String (well, a PathBuf, but invalid UTF-8 is a
hard error, so it's basically a String but worse, but that's a separate
issue) while the other variants are tables with a "type" field acting as
a discriminant.

This means another method is needed, such as the existing custom
Deserialize impl, but it's a bit uglier than necessary and in particular
would have to be duplicated in order to provide a Serialize impl as
well.

I created a pair of secondary enums - an outer one using untagged
representation and an inner one with internally tagged representation -
added some From impls and annotated FileTarget itself with #[serde(from,
into)], so Serde's derive macros do the heavy lifting instead of a
custom impl. There is still some boilerplate in the From impls, but it's
much smaller and cleaner now.