Move path canonicalization
Call real_path in filesystem::compare_symlink instead of at its
call-sites in filesystem::RealFilesystem::compare_symlink and
filesystem::DryRunFilesystem::compare_symlink, to avoid real_path being
called unnecessarily and in particular when the path refers to a
non-existent file, which would cause an unhandled error.

This error would occur when deleting a file from within a source
directory and running dotter deploy; instead of deleting the
corresponding (now-broken) symlink, dotter would crash.

This is not the only cause for this crash, unfortunately.

I have not tested the changes to Windows code, but the only difference
is a pair of logging macros, so I can't see this causing problems.