Don't follow links in remove_file()
Use symlink_metadata() instead of metadata() to avoid errors when the
file in question is a broken symlink.

If path points to a symlink, the is_dir() check would return whether the
link target is a directory, rather than returning false (as symlinks are
not dirs), but thankfully remove_dir_all removes the link rather than
recursively removing the link target in this case. I'm not certain this
could actually happen in practice though, as Dotter generally checks
things before calling remove_file() and does not currently create
symlinks to directories.