add options for disabling recursion
This adds an option for dealing with both directory sources and
symlinks. When `recurse` is set to false, a source is not traversed
until it lands on a regular file; it stops there and makes the source
directory/symlink the target of the destination symlink. This option can
be set per `SymbolicTarget` to retain the default behavior for other
packages, or it can be set at the global level to change the default
behavior, i.e. it makes the sources and destinations "what you see is
what you get", for those who might prefer this behavior. It retains the
default behavior to avoid breakage.