Run hook scripts directly from cache directory
Copying the hooks to temp is a problem if you run dotter with different
users, as the file gets left behind and other users don't have
permissions to overwrite it. It's also not really needed to copy the
file to somewhere else, it can just be run to where it is anyway from
the cache directory.

It might be also a problem to create the hooks world readable at the
temp-directory, where other users can read it, it might contain secrets
that aren't expected to be written to outside of the home directory.