Allow passing variables defined by `assign` to `include_template`.
This change allows using the `assign` helper from
`handlebars_misc_helpers` to pass variables into templates rendered with
`include_template`. The `assign` helper modifies the `RenderContext` and
previously `include_template` would use the `Context` when rendering the
template. The change attempts to use the `RenderContext` with a fallback
to use the `Context` if no `RenderContext` exists.

Note that this change does not address `handlebars_misc_helpers` #65.
The `assign` helper clones the `Context` and overwrites the
`RenderContext` on every call meaning that only variables defined by the
most recent `assign` call will be available.

Resolves #134.